requirements.txt
paddlenlp=2.3.4
paddlepaddle-gpu=2.3.0
